requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.3.0
enum34==1.1.6
gitdb2==2.0.5
GitPython==2.1.11
ipython==5.8.0
ipython-genutils==0.2.0
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
scandir==1.9.0
simplegeneric==0.8.1
six==1.11.0
smmap2==2.0.5
traitlets==4.3.2
wcwidth==0.1.7
